+++ ce/main.ts
@@ -137,6 +137,7 @@ async function main() {
 
     await session.writePostscript();
   } catch (e) {
+    console.log((<Error>e).stack);
     error(e);
     return process.exitCode = 1;
   } finally {
